requirements.txt
coreapi==2.0.9
Django==1.10.3
django-debug-toolbar==1.6
django-filter==0.15.3
django-rest-swagger==2.1.0
djangorestframework==3.5.3
gunicorn==19.6.0
itypes==1.1.0
Markdown==2.6.7
openapi-codec==1.1.7
PyMySQL==0.7.9
pytz==2016.7
requests==2.11.1
simplejson==3.10.0
sqlparse==0.2.2
uritemplate==3.0.0
